requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch>=2.6.0
torchvision>=0.17.0
Pillow==10.1.0
gTTS==2.4.0
python-multipart==0.0.6 